requirements.txt
# GeminiCRM Pro - Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
flask-limiter>=3.5.0

# Authentication & Security
PyJWT>=2.8.0
cryptography>=41.0.0

# Google Gemini AI
google-genai>=1.0.0

# Database & ORM (for production)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Email & Notifications
python-dateutil>=2.8.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
python-slugify>=8.0.0

# Performance & Monitoring
redis>=5.0.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0

